[Add] 플레이어 이동 및 조작, 글라이딩 구현

+++ Assets/01.Scripts/Combat/DamageInfo.cs
@@ -2,11 +2,11 @@ using UnityEngine;
 
 public readonly struct DamageInfo
 {
-    public readonly int Amount;
+    public readonly float Amount;
     public readonly Vector2 HitPoint;
     public readonly Vector2 HitDirection;
 
-    public DamageInfo(int amount, Vector2 hitPoint, Vector2 hitDirection)
+    public DamageInfo(float amount, Vector2 hitPoint, Vector2 hitDirection)
     {
         Amount = amount;
         HitPoint = hitPoint;
